refactor: change component file name to lowercase
